Brings the MIDI subsystem work onto main alongside the in-flight tutti-core
simplification (PDC rework + MeteringManager split).

The two lines of work turned out to be the same refactor applied to different
concerns, which is what made the conflicts mechanical rather than semantic:
main pulled `pdc` out of `AudioGraph`, this branch pulled `midi_route` out.
Every conflict was the two sides deleting a different field from the same
struct, signature, or helper — so the resolution keeps NEITHER:

- `AudioGraph` now holds only `net` / `sample_rate` / `channels`.
- `from_parts(net, sample_rate, channels)` — both extra params gone.
- `commit()` returns `()` (main's PDC-latency return dropped) and no longer
  publishes a routing snapshot; latency compensation and MIDI routing each
  publish themselves now.
- The three `bare_graph` test helpers in graph/{reconcile,routing,sidechain}
  are deleted (main's side): with nothing extra to construct, `AudioGraph::empty`
  covers them, and main had already removed every caller.
- `commit_publishes_per_channel_latency` in editable.rs is dropped — it tested
  `pdc_snapshot()` and the `usize` return, neither of which exists after main's
  PDC rework. Deleted with the feature, not worked around.

The module doc on `AudioGraph` now explains both departures together, since
they share a rationale: neither latency nor MIDI-channel routing is a property
of the DSP graph.

Verified on the merged tree: tutti-core 113+3+2+1, tutti-midi-types 161+3+2,
tutti-midi-runtime 99+3, tutti-midi-io 44+1 doc-test; clippy -D warnings clean
across tutti-core / tutti-midi-io / tutti-midi-runtime; `cargo check --bin dawai`
and the no-default-features bevy-tutti build both clean. (tutti-core's count
drops from 137 because main's PDC + metering deletions removed those tests.)

+++ crates/midi/tutti-midi-types/src/ump/utility.rs
@@ -58,7 +58,7 @@ mod tests {
         let ev = MidiEvent::jr_timestamp(0, 0x1234);
         match UmpMessage::try_from(ev.data_words()).unwrap() {
             UmpMessage::Utility(utility::Utility::Timestamp(m)) => {
-                assert_eq!(u16::from(m.time_data()), 0x1234);
+                assert_eq!(m.time_data(), 0x1234);
             }
             other => panic!("expected JR Timestamp, got {other:?}"),
         }
